encapsulate components - Subset
