+++ src/components/Subset/index.js
@@ -1,0 +1,4 @@
+export * from './Subset';
+export * from './SubsetBrief';
+export * from './Subsets';
+export * from './SubsetsBanner';
